finish class reorganization; move most code into watertap_costing_package module
